+++ docs/tutorials/mlflow.rst
@@ -5,7 +5,7 @@ In this post, we are going to explain how setup the build-in integration
 of sklearn-genetic-opt with MLflow.
 To use this feature, we must set the parameters that will include
 the tracking server, experiment name, run name, tags and others,
-the full implementation is here: :class:`~sklearn_genetic.mlflow.MLflowConfig`
+the full implementation is here: :class:`~sklearn_genetic.mlflow_log.MLflowConfig`
 
 Configuration
 -------------
@@ -29,7 +29,7 @@ trained models.
 
 .. code:: python3
 
-    from sklearn_genetic.mlflow import MLflowConfig
+    from sklearn_genetic.mlflow_log import MLflowConfig
 
     mlflow_config = MLflowConfig(
         tracking_uri="http://localhost:5000",
